Feature/pool 128 the prize strategy award arbitrary tokens (#51)
* PrizeStrategy can award arbitrary tokens

- Added ability for the PrizeStrategy to award arbitrary tokens held by
the PrizePool
- cTokens are restricted for the CompoundPrizePool

* Add solhint, organize dependencies

* Refactors as per review, code clean-up as per solhint

* Fix require statements

* Improved test coverage
