+++ .solhint.json
@@ -1,0 +1,18 @@
+{
+  "extends": "solhint:recommended",
+  "plugins": [],
+  "rules": {
+      "func-order": "off",
+      "mark-callable-contracts": "off",
+      "no-empty-blocks": "off",
+      "compiler-version": ["error", "^0.6.0"],
+      "private-vars-leading-underscore": "off",
+      "code-complexity": "warn",
+      "const-name-snakecase": "warn",
+      "function-max-lines": "warn",
+      "max-line-length": ["warn", 160],
+      "avoid-suicide": "error",
+      "avoid-sha3": "warn",
+      "not-rely-on-time": "off"
+  }
+}
