+++ contracts/external/openzeppelin/ProxyFactory.sol
@@ -21,7 +21,7 @@ contract ProxyFactory {
 
     if(_data.length > 0) {
       (bool success,) = proxy.call(_data);
-      require(success, "constructor call failed");
+      require(success, "ProxyFactory/constructor-call-failed");
     }
   }
 }
